==> notebooks/1. get player match ID's from dotabuff.com.ipynb <==
#Scrape match ID's from dotabuff.com, since developers' Web API
#does not have the option to retrieve all of player's matches easily. 
import requests
from bs4 import BeautifulSoup as bs
import pickle

player_ID = #Your player ID goes here
match_ID_list = []
page = 1

url = "https://www.dotabuff.com/players/{player_ID}/matches?enhance=overview&page={page}"

while True:
    r   = requests.get(url.format(player_ID = player_ID, page=page), headers={'user-agent': 'Mozilla/5.0'})
    html = r.text
    my_soup = bs(html, 'html.parser')
    
    match_data_per_page = my_soup.find_all('td', class_ = "cell-large", attrs={'href'})
    
    for td in match_data_per_page:
        for a in td.find_all('a'):
            temp = a.get('href')
            match_ID_list.append(int(temp[9:]))

    if my_soup.find(class_="next") is None:
            break  
            
    page = page + 1

#Save results on your computer           
pickle_out = open("match_ID_list.pickle","wb")
pickle.dump(match_ID_list, pickle_out)
pickle_out.close()
