==> notebooks/2. auxiliary item dictionary to convert numeric item ID's to their names.ipynb <==
import json
import pickle

#Working with the dictionary provided in https://github.com/joshuaduffy/dota2api
with open('../../dota2api-master/dota2api/ref/items.json') as json_data_file:
    item_ref = json.load(json_data_file)

item_dictionary = {}
x = 0

for items in item_ref['items']:
    item_dictionary[item_ref['items'][x]['id']] = item_ref['items'][x]['localized_name']
    x = x + 1
    
#Delete seasonal cosmetic items
del item_dictionary[1021]
del item_dictionary[1022]
del item_dictionary[1023]
del item_dictionary[1024]
del item_dictionary[1025]
del item_dictionary[1026]
del item_dictionary[1027]

#Add "none" for items which are 0
item_dictionary[0] = 'None'

#Save the result on your computer           
pickle_out = open("item_dictionary.pickle","wb")
pickle.dump(item_dictionary, pickle_out)
pickle_out.close()
