==> notebooks/4. preliminary data pre-processing.ipynb <==
import pickle
import pandas as pd
import numpy as np
from IPython.display import display
pd.options.display.max_columns = None
pd.options.display.max_rows = None

pickle_in = open("match_dataframe.pickle","rb")
match_dataframe = pickle.load(pickle_in)

display(match_dataframe)

#Convert timestamps into standard date format
match_dataframe['Date'] = pd.to_datetime(match_dataframe.Date, unit = 's')

#Delete LobbyName column, since it is simply a repeat of LobbyType column
del match_dataframe['LobbyName']

#Convert match duration from seconds to HH:MM:SS format
match_dataframe['Duration'] = pd.to_timedelta(match_dataframe['Duration'], unit='s')

#Check Lobby Types
match_dataframe['LobbyType'].unique()

#Remove all rows with LobbyType 4, since these are bot matches
match_dataframe = match_dataframe[match_dataframe.LobbyType != 4]

#Check Game Mode
match_dataframe['GameMode'].unique()
match_dataframe['GameMode'].value_counts()

#Since this account does not have any ranked matches (Game Mode #22), there must be an error on the storing side. 
#So I am going to replace all values of 22 with 1.
match_dataframe['GameMode'].replace(to_replace = 22, value = 1, inplace = True)

#Remove all values that are not 1
match_dataframe = match_dataframe[match_dataframe.GameMode == 1]

#Now that I have a certain kind of matches which I want to analyze, I can just remove
#LobbyType and GameMode columns from the dataframe
del match_dataframe['LobbyType']
del match_dataframe['GameMode']

#Check Leaver Status column
match_dataframe['LeaverStatus'].unique()
match_dataframe['LeaverStatus'].value_counts()

#Around 18% of the games have leavers, but since there is an equal chance of leavers being on either team,
#I will ignore this statistics and keep these 18% of data in the dataframe.

#Store the result            
pickle_out = open("match_dataframe_processed.pickle","wb")
pickle.dump(match_dataframe, pickle_out)
pickle_out.close()
