==> notebooks/6. task 2 - item analysis.ipynb <==
import pickle
import pandas as pd
import numpy as np
from IPython.display import display
pd.options.display.max_columns = None
pd.options.display.max_rows = None

pickle_in = open("match_dataframe_processed.pickle","rb")
match_dataframe = pickle.load(pickle_in)

display(match_dataframe)

#I want to analyze my item selection in the games which I have won to draw some insights into what items
#give me the best chance at winning. To this end, I am going to look at items in general and then create a dataframe
#with only won matches to see the item calculations.

#Unfortunately, there is no easy way to do it given the structure of the dataframe, so I will do it with dictionaries.
Item1_dict = match_dataframe.groupby('Hero')['Item1'].value_counts().to_dict()
Item2_dict = match_dataframe.groupby('Hero')['Item2'].value_counts().to_dict()
Item3_dict = match_dataframe.groupby('Hero')['Item3'].value_counts().to_dict()
Item4_dict = match_dataframe.groupby('Hero')['Item4'].value_counts().to_dict()
Item5_dict = match_dataframe.groupby('Hero')['Item5'].value_counts().to_dict()
Item6_dict = match_dataframe.groupby('Hero')['Item6'].value_counts().to_dict()

item_dicts = [Item1_dict, Item2_dict, Item3_dict, Item4_dict, Item5_dict, Item6_dict]

result = {}
for my_dict in item_dicts:
    for key, value in my_dict.items():
        result.setdefault(key, 0)
        result[key] = result[key] + value
        
print(result)

#This result is messy, so what if I sort it?
sorted(result, key=result.get)

for item in sorted(result, key=result.get, reverse=True):
    print(item, result[item])
    
#This is still hard to read, so let's unstack this dictionary and turn it into a dataframe:
item_dataframe = pd.DataFrame()

item_dataframe = pd.Series(result)
item_dataframe = item_dataframe.unstack()

display(item_dataframe)

#This is better, but what if I wanted to look at specific heroes?
hero = #Desired hero's name goes here

item_dataframe = item_dataframe.transpose()
display(item_dataframe[hero].dropna().drop('None').sort_values(ascending = False))

#Now, these are the totals. I want to see the percentage of these items winning.
#To this end, I am going to store the results in a dictionary...
hero_item_dict = item_dataframe[hero].dropna().drop('None').to_dict()
print(hero_item_dict)

#...and repeat the process for Won matches:
match_dataframe_won = match_dataframe[match_dataframe.Won == 1].copy()

display(match_dataframe_won)

#Do the same manipulations: 
Item1_dict_won = match_dataframe_won.groupby('Hero')['Item1'].value_counts().to_dict()
Item2_dict_won = match_dataframe_won.groupby('Hero')['Item2'].value_counts().to_dict()
Item3_dict_won = match_dataframe_won.groupby('Hero')['Item3'].value_counts().to_dict()
Item4_dict_won = match_dataframe_won.groupby('Hero')['Item4'].value_counts().to_dict()
Item5_dict_won = match_dataframe_won.groupby('Hero')['Item5'].value_counts().to_dict()
Item6_dict_won = match_dataframe_won.groupby('Hero')['Item6'].value_counts().to_dict()

item_dicts_won = [Item1_dict_won, Item2_dict_won, Item3_dict_won, Item4_dict_won, Item5_dict_won, Item6_dict_won]

result_won = {}

for my_dict in item_dicts_won:
    for key, value in my_dict.items():
        result_won.setdefault(key, 0)
        result_won[key] = result_won[key] + value
        
print(result_won)

#And again...
item_dataframe_won = pd.DataFrame()

item_dataframe_won = pd.Series(result_won)
item_dataframe_won = item_dataframe_won.unstack()

display(item_dataframe_won)

item_dataframe_won = item_dataframe_won.transpose()

display(item_dataframe_won[hero].dropna().drop('None').sort_values(ascending = False))

#Now, these are the totals. I want to see the percentage of these items winning, so I am going to do this
hero_item_dict_won = item_dataframe_won[hero].dropna().drop('None').to_dict()
print(hero_item_dict_won)

print(len(hero_item_dict))
print(len(hero_item_dict_won))

#Makes sense: there are more items in total than items with which I won my games
for item in hero_item_dict:
    try:
        print(item, ": ", round(hero_item_dict_won[item]/hero_item_dict[item], 2))
    except KeyError:
        print(item, ": 0")
        
#Now, calculate item win percentage and put it in a dictionary:
win_percentage = {}
for item in hero_item_dict:
    try:
        win_percentage[item] = round(hero_item_dict_won[item]/hero_item_dict[item], 2)
    except KeyError:
        win_percentage[item] = 0
        
for item in sorted(win_percentage, key=win_percentage.get, reverse=True):
    print(item, win_percentage[item])
    
#That's not that helpful, so let's take a look at the top 5 bought items from item_dataframe: 
#Eul's Scepter of Divinity, Aghanim's Scepter, Soul Ring, Arcane Boots, Boots of Travel, Bloodstone, Force Staff:
item1 = #desired item goes here
item2 = #desired item goes here
item3 = #desired item goes here
item4 = #desired item goes here
item5 = #desired item goes here

print(win_percentage[item1])
print(win_percentage[item2])
print(win_percentage[item3])
print(win_percentage[item4])
print(win_percentage[item5])

#This gives an idea which buying which items will help me win the game.
